LPS-93516 Move integration tests of com.liferay.portal.service.user package to module
