+++ modules/apps/user/user-test/src/testIntegration/java/com/liferay/user/service/test/UserServiceTestUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.service.user;
+package com.liferay.user.service.test;
 
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
