+++ modules/apps/user/user-test/src/testIntegration/java/com/liferay/user/service/test/UserServiceWhenCompanySecurityStrangersWithMXDisabledTest.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.service.user;
+package com.liferay.user.service.test;
 
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.portal.kernel.exception.UserEmailAddressException;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.security.auth.PrincipalThreadLocal;
@@ -30,12 +31,14 @@ import com.liferay.portal.util.PropsUtil;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 /**
  * @author Brian Wing Shun Chan
  * @author José Manuel Navarro
  * @author Drew Brokke
  */
+@RunWith(Arquillian.class)
 public class UserServiceWhenCompanySecurityStrangersWithMXDisabledTest {
 
 	@ClassRule
